Added error handling, closed / open trades filter, layout improvments, created a common TradeMapper, added invested capital

+++ trade-journal-backend/src/main/java/de/kaluza/trades/domains/SellTradeEntity.java
@@ -41,7 +41,7 @@ public class SellTradeEntity {
     private double performanceInPercent;
 
     @Column(nullable = false, updatable = false)
-    private long tradingDays;
+    private int tradingDays;
 
     @ManyToOne(optional = false, cascade = CascadeType.ALL)
     @JoinColumn(
